+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * revise gemspec [jcupitt]
 * allow symbols for Interpolate.new [noraj]
 * update docs for 8.10, fix minor doc formatting issues [jcupitt]
+* add "mutate" method [jcupitt]
 
 ## Version 2.0.17 (2019-10-29)